Move to api? This might get reverted later if we find a better place
